♻️ Refactor : Service에 정의되어 있는 오류 메세지 정의를 Repository로 옮기는 작업
