+++ src/main/java/com/example/codingmall/Order/OrderService.java
@@ -29,15 +29,16 @@ public class OrderService {
                 .map(orderItemRequest -> {
                     Item item = itemRepository.findItemById(orderItemRequest.getItemId());
                     //item.removeStock(orderItemRequest.getItemCount()); --> 일단 테스트를 위해 잠시만 꺼놓음.
-                    return OrderItem.createOrderItem(item, orderItemRequest.getItemCount());  // null은 orderId가 없으므로 임시로 처리
+                    return OrderItem.createOrderItem(item, orderItemRequest.getItemCount());
                 })
                 .collect(Collectors.toList());
+
         // Order 객체를 생성 (totalAmount는 createOrder에서 자동으로 계산됨)
         Order order = Order.createOrder(user, orderRequest, orderItems);
 
         // Order의 OrderItem을 연결
         for (OrderItem orderItem : orderItems) {
-            orderItem.setOrder(order);  // 각 OrderItem에 해당하는 Order를 연결
+            orderItem.setOrder(order);
         }
 
         // 주문 저장
@@ -48,7 +49,8 @@ public class OrderService {
     /* 장바구니 상품들 주문 */
     @Transactional
     public Long createOrderFromCart(User user, OrderCartRequest orderCartRequest) {
-        Cart cart = cartRepository.findByUser(user).orElseThrow(() -> new IllegalStateException("유저의 장바구니를 찾지 못 했습니다."));
+        Cart cart = cartRepository.findCartByUser(user);
+
         List<OrderItem> orderItems = cart.getItems().stream() //Cart 안의 CartItem을 순환
                 .map(cartItem -> {
                     Item item = cartItem.getItem();
@@ -56,17 +58,19 @@ public class OrderService {
                     return OrderItem.createOrderItem(item, cartItem.getCount());
                 })
                 .collect(Collectors.toList());
+
         OrderRequest orderRequest1 = OrderRequest.builder()
                 .receiverName(orderCartRequest.getReceiverName())
                 .receiverPhone(orderCartRequest.getReceiverPhone())
                 .deliveryAddress(orderCartRequest.getDeliveryAddress())
                 .orderNote(orderCartRequest.getOrderNote())
                 .build();
+
         Order order = Order.createOrder(user, orderRequest1, orderItems);
 
         // Order의 OrderItem을 연결
         for (OrderItem orderItem : orderItems) {
-            orderItem.setOrder(order);  // 각 OrderItem에 해당하는 Order를 연결
+            orderItem.setOrder(order);
         }
 
         orderRepository.save(order);
@@ -77,7 +81,7 @@ public class OrderService {
     /* 주문 취소 */
     @Transactional
     public void cancelOrder(Long orderId) {
-        Order order = orderRepository.findById(orderId).orElseThrow(() -> new IllegalArgumentException("주문을 찾지 못했습니다."));
+        Order order = orderRepository.findOrderById(orderId);
         order.cancel();
     }
 }
