+++ src/main/java/com/example/codingmall/User/UserService.java
@@ -15,8 +15,7 @@ public class UserService implements UserDetailsService {
     private final UserRepository userRepository;
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        User user = userRepository.findByUsername(username)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found with username: " + username));
+        User user = userRepository.findUserByUsername(username);
         return user;
     }
 }
